fix(bot): require Feishu webhook token; split weixin account/login out (#3779)
Follow-up to #3763.

- Feishu webhook mode is the default, and verificationTokenValid accepts
  any caller when no token is configured, so an empty verification_token
  exposed an unauthenticated event endpoint. Refuse to start webhook mode
  without a token (fail closed).
- Move the weixin account-persistence and QR-login block to
  weixin_login.go, bringing weixin.go under the 800-line ceiling (811->587).
